chore(build): rework workflow files
The commit removes the lint job from the GitHub workflow for npm-publish and refactors the test job to use a separate workflow. A notable change is the removal of the configuration that sets up older Node versions (< 18). Consequently it should be expected that support for those versions will drop. The Node version setup and 'check-latest' directive have also been removed from the Github registry package (gpr) publish workflow. These changes aim to streamline the workflows, reducing redundancy and ensuring tests run on Node versions that are actively supported.